demo/requirements.txt
aniso8601==8.0.0
asgiref==3.2.3
attrs==19.3.0
awscli==1.18.15
awsebcli==3.17.1
bcrypt==3.1.7
blessed==1.17.2
botocore==1.14.17
cached-property==1.5.1
cement==2.8.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.3.9
configparser==4.0.2
cryptography==2.8
Django==3.0.4
docker==4.2.0
docker-compose==1.25.4
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
ffmpeg==1.4
ffmpeg-python==0.2.0
ffmpy==0.2.2
Flask==1.1.1
Flask-RESTful==0.3.8
future==0.16.0
idna==2.7
image==1.5.28
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
jsonschema==3.2.0
MarkupSafe==1.1.1
more-itertools==8.2.0
oauthlib==3.1.0
packaging==20.3
paramiko==2.7.1
pathspec==0.5.9
Pillow==7.0.0
pluggy==0.13.1
py==1.8.1
pyasn1==0.4.8
pycparser==2.20
PyNaCl==1.3.0
pyparsing==2.4.6
pyrsistent==0.15.7
PySocks==1.7.1
pytest==5.3.5
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.2
requests==2.20.1
requests-oauthlib==1.3.0
rsa==3.4.2
s3transfer==0.3.3
semantic-version==2.5.0
six==1.11.0
sqlparse==0.3.1
termcolor==1.1.0
texttable==1.6.2
tweepy==3.8.0
urllib3==1.24.3
wcwidth==0.1.8
websocket-client==0.57.0
Werkzeug==1.0.0
WTForms==2.2.1
zipp==3.1.0
